requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.14.0
numpy>=1.24.0
tqdm>=4.65.0
accelerate>=0.20.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
peft>=0.7.0
